# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AllocationResponse:
    properties:
      aggregates:
        additionalProperties:
          $ref: '#/definitions/dto.AggregateResponse'
        type: object
      aggregates_timeline:
        items:
          additionalProperties:
            $ref: '#/definitions/dto.AggregateResponse'
          type: object
        type: array
      allocation_targets:
        items:
          $ref: '#/definitions/dto.AllocationTargetResponse'
        type: array
    type: object
...
  dto.AssetsBalanceResponse:
    properties:
      asset_breakdowns:
        additionalProperties:
          $ref: '#/definitions/dto.AssetBreakdownResponse'
        type: object
    type: object
...
  dto.TfIdfResponse:
    properties:
      index:
        $ref: '#/definitions/dto.IndexResponse'
      tf_idf:
        additionalProperties:
          additionalProperties:
            format: float64
            type: number
          type: object
        type: object
    type: object
...
  server.LogsResponse:
    properties:
      logs:
        items: {}
        type: array
    type: object
...
info:
  contact: {}
  description: Paisa is a local-first personal finance application. This API powers
    the Paisa frontend and exposes financial analysis, journal editing, configuration,
    price data, goals, transactions, and related application operations.
  title: Paisa API
  version: 0.1.0
paths:
  /account/tf_idf:
    get:
      description: Returns TF-IDF prediction model data
      operationId: getTfIdf
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TfIdfResponse'
      security:
      - PaisaAuth: []
      summary: Get TF-IDF machine learning model for payee-account prediction
      tags:
      - Predictions
  /allocation:
    get:
      description: Returns asset allocation targets and current asset distribution
      operationId: getAllocation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AllocationResponse'
      security:
      - PaisaAuth: []
      summary: Get asset class allocations
      tags:
      - Allocation
  /assets/balance:
    get:
      description: Returns balance breakdown for asset accounts
      operationId: getAssetsBalance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AssetsBalanceResponse'
      security:
      - PaisaAuth: []
      summary: Get current asset balances and breakdowns
      tags:
      - Assets
  /budget:
    get:
      description: Returns monthly budget allocations, spending actuals, variances,
        and rollover balances
      operationId: getBudget
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.BudgetsSummaryResponse'
      security:
      - PaisaAuth: []
      summary: Get budget forecasts and actuals
      tags:
      - Budget
  /capital_gains:
    get:
      description: Computes FIFO realized capital gains grouped by financial year
      operationId: getCapitalGains
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CapitalGainsResponse'
      security:
      - PaisaAuth: []
      summary: Get FIFO realized capital gains
      tags:
      - Tax
  /cash_flow:
    get:
      description: Returns monthly cash flow breakdown including inflows, outflows,
        and net changes
      operationId: getCashFlow
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CashFlowsResponse'
      security:
      - PaisaAuth: []
      summary: Get monthly cash flow statement
      tags:
      - Cash Flow
  /config:
    get:
      description: Returns public configuration, account list, current time override,
        and JSON validation schema
      operationId: getConfig
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PublicConfigResponse'
      security:
      - PaisaAuth: []
      summary: Get application configuration and metadata
      tags:
      - Configuration
    post:
      consumes:
      - application/json
      - text/plain
      description: Overwrites the paisa.yaml configuration file. No-op in readonly
        mode.
      operationId: saveConfig
      parameters:
      - description: Configuration YAML text
        in: body
        name: config
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Save configuration YAML
      tags:
      - Configuration
  /credit_cards:
    get:
      description: Returns credit card balances, due dates, and statements
      operationId: getCreditCards
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreditCardsResponse'
      security:
      - PaisaAuth: []
      summary: Get all credit cards summaries and billing cycles
      tags:
      - Credit Cards
  /credit_cards/{account}:
    get:
      description: Returns credit card details for a single account
      operationId: getCreditCard
      parameters:
      - description: Credit card account name (e.g. Liabilities:CreditCard:AmazonPay)
        in: path
        name: account
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreditCardSummaryResponse'
      security:
      - PaisaAuth: []
      summary: Get credit card summary for a specific account
      tags:
      - Credit Cards
  /dashboard:
    get:
      description: Returns consolidated KPI summaries across net worth, expenses,
        budgets, and investments
      operationId: getDashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DashboardResponse'
      security:
      - PaisaAuth: []
      summary: Get dashboard financial summary
      tags:
      - Dashboard
  /diagnosis:
    get:
      description: Runs system diagnostics and returns detected issues
      operationId: getDiagnosis
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DiagnosisResponse'
      security:
      - PaisaAuth: []
      summary: Run system diagnostic health checks
      tags:
      - Diagnosis
  /editor/file:
    post:
      consumes:
      - application/json
      description: Returns contents and version history of a ledger file
      operationId: getEditorFile
      parameters:
      - description: Ledger file query
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/dto.LedgerFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LedgerFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Read a specific ledger file and its backup versions
      tags:
      - Editor
  /editor/file/delete_backups:
    post:
      consumes:
      - application/json
      description: Deletes all timestamped backup files for a ledger file
      operationId: deleteEditorBackups
      parameters:
      - description: Ledger file
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/dto.LedgerFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LedgerFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Delete backup versions for a ledger file
      tags:
      - Editor
  /editor/files:
    get:
      description: Returns all ledger files and their current postings
      operationId: getEditorFiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorFilesResponse'
      security:
      - PaisaAuth: []
      summary: List all ledger files in journal directory
      tags:
      - Editor
  /editor/save:
    post:
      consumes:
      - application/json
      description: Atomically writes ledger file with automatic timestamped backup.
        No-op in readonly mode.
      operationId: saveEditorFile
      parameters:
      - description: Ledger file content
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/dto.LedgerFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorSaveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Save ledger file and trigger database synchronization
      tags:
      - Editor
  /editor/validate:
    post:
      consumes:
      - application/json
      description: Parses and validates ledger file syntax without saving
      operationId: validateEditorFile
      parameters:
      - description: Ledger file to validate
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/dto.LedgerFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EditorValidateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Validate ledger file syntax
      tags:
      - Editor
  /expense:
    get:
      description: Computes hierarchical flow graph and periodic expense summaries
      operationId: getExpense
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExpenseResponse'
      security:
      - PaisaAuth: []
      summary: Get expense hierarchy graph and breakdown
      tags:
      - Expenses
  /gain:
    get:
      description: Computes realized, unrealized gains and annualized XIRR returns
        across accounts
      operationId: getGain
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GainsResponse'
      security:
      - PaisaAuth: []
      summary: Get investment gains and XIRR performance
      tags:
      - Gains
  /gain/{account}:
    get:
      description: Computes realized and unrealized gain for a single investment account
      operationId: getAccountGain
      parameters:
      - description: Account name (e.g. Assets:Equity:Nifty)
        in: path
        name: account
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AccountGainResponse'
      security:
      - PaisaAuth: []
      summary: Get investment gain for a specific account
      tags:
      - Gains
  /goals:
    get:
      description: Returns summary and progress percentage for all configured goals
      operationId: getGoals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GoalSummariesResponse'
      security:
      - PaisaAuth: []
      summary: List savings and retirement goals summaries
      tags:
      - Goals
  /goals/{type}/{name}:
    get:
      description: Returns progress, projections, and monthly timeline for a single
        goal
      operationId: getGoalDetails
      parameters:
      - description: Goal type (savings or retirement)
        in: path
        name: type
        required: true
        type: string
      - description: Goal name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GoalDetailResponse'
      security:
      - PaisaAuth: []
      summary: Get details for a specific goal
      tags:
      - Goals
  /harvest:
    get:
      description: Returns tax harvesting opportunities and current gains
      operationId: getHarvest
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HarvestResponse'
      security:
      - PaisaAuth: []
      summary: Get tax harvesting opportunities
      tags:
      - Tax
  /income:
    get:
      description: Returns gross income, taxes, net income, and periodic income timelines
      operationId: getIncome
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncomeResponse'
      security:
      - PaisaAuth: []
      summary: Get income timeline and yearly summary
      tags:
      - Income
  /income_statement:
    get:
      description: Returns yearly income statements with income, expenses, interest,
        and taxes
      operationId: getIncomeStatement
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncomeStatementResponse'
      security:
      - PaisaAuth: []
      summary: Get financial-year income statements
      tags:
      - Income Statement
  /init:
    post:
      description: Generates demo ledger data and configuration. No-op in readonly
        mode.
      operationId: initDemoData
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Initialize demo data
      tags:
      - Initialization
  /insights:
    get:
      description: Returns deterministic financial observations and risks derived
        from ledger history
      operationId: getInsights
      parameters:
      - description: Month period in YYYY-MM format (defaults to current month)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InsightsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Get deterministic financial health insights
      tags:
      - Insights
  /investment:
    get:
      description: Returns yearly investment cards and savings rate metrics
      operationId: getInvestment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.InvestmentResponse'
      security:
      - PaisaAuth: []
      summary: Get investment summary and savings rate
      tags:
      - Investments
  /ledger:
    get:
      description: Returns flat array of all journal postings with computed balances
        and market values
      operationId: getLedger
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PostingsResponse'
      security:
      - PaisaAuth: []
      summary: Get all ledger postings with market valuation
      tags:
      - Ledger
  /liabilities/balance:
    get:
      description: Computes outstanding loan balances
      operationId: getLiabilitiesBalance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LiabilitiesBalanceResponse'
      security:
      - PaisaAuth: []
      summary: Get loan balances and breakdowns
      tags:
      - Liabilities
  /liabilities/interest:
    get:
      description: Computes interest payments and effective APR across loans
      operationId: getLiabilitiesInterest
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LiabilitiesInterestResponse'
      security:
      - PaisaAuth: []
      summary: Get loan interest and APR calculations
      tags:
      - Liabilities
  /liabilities/repayment:
    get:
      description: Returns loan repayment postings
      operationId: getLiabilitiesRepayment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LiabilitiesRepaymentResponse'
      security:
      - PaisaAuth: []
      summary: Get loan repayment history
      tags:
      - Liabilities
  /logs:
    get:
      description: Returns recent application log lines
      operationId: getLogs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.LogsResponse'
      security:
      - PaisaAuth: []
      summary: Get application log entries
      tags:
      - Logs
  /networth:
    get:
      description: Returns the current net-worth analysis and historical timeline
      operationId: getNetworth
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.NetworthResponse'
      security:
      - PaisaAuth: []
      summary: Get net worth
      tags:
      - Net Worth
...
  /portfolio_allocation:
    get:
      description: Returns asset allocations grouped by portfolio
      operationId: getPortfolioAllocation
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PortfolioAllocationResponse'
      security:
      - PaisaAuth: []
      summary: Get portfolio-grouped allocations
      tags:
      - Allocation
  /prediction/history:
    get:
      description: Returns prediction history records
      operationId: getPredictionHistory
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PredictionHistoryResponse'
      security:
      - PaisaAuth: []
      summary: Get prediction history
      tags:
      - Predictions
  /price:
    get:
      description: Returns price histories grouped by commodity
      operationId: getPrices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PricesResponse'
      security:
      - PaisaAuth: []
      summary: Get all cached commodity prices
      tags:
      - Prices
  /price/autocomplete:
    post:
      consumes:
      - application/json
      description: Searches ticker symbols from configured scrapers
      operationId: getPriceAutoCompletions
      parameters:
      - description: Autocomplete query
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/server.AutoCompleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AutoCompleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - PaisaAuth: []
      summary: Autocomplete commodity or ticker symbols
      tags:
      - Prices
  /price/delete:
    post:
      description: Clears in-memory and database cached commodity prices. No-op in
        readonly mode.
      operationId: clearPriceCache
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
      security:
      - PaisaAuth: []
      summary: Clear cached market prices
      tags:
      - Prices
  /price/providers:
    get:
      description: Returns registered price scraping providers and their sync status
      operationId: getPriceProviders
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PriceProvidersResponse'
      security:
      - PaisaAuth: []
      summary: Get configured price providers
      tags:
      - Prices
  /price/providers/delete/{provider}:
    post:
      description: Clears cached commodity prices for a specific provider. No-op in
        readonly mode.
      operationId: clearPriceProviderCache
      parameters:
      - description: Provider code (e.g. yahoo)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
      security:
      - PaisaAuth: []
      summary: Clear price cache for a specific provider
      tags:
      - Prices
  /recurring:
    get:
      description: Returns recurring monthly or periodic transaction sequences
      operationId: getRecurringTransactions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RecurringTransactionsResponse'
      security:
      - PaisaAuth: []
      summary: Get recurring transaction sequences
      tags:
      - Recurring
...
  /sheets/files:
    get:
      description: Returns list of .paisa sheet files
      operationId: getSheetFiles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SheetsResponse'
      security:
      - PaisaAuth: []
      summary: List all .paisa sheet query files
      tags:
      - Sheets
...
securityDefinitions:
  PaisaAuth:
    description: Credentials formatted as 'username:password'. Required for all /api
      endpoints when user accounts are configured in paisa.yaml.
    in: header
    name: X-Auth
    type: apiKey
swagger: "2.0"